[type-sig] fix fatal when lib contains toplevel import
Summary:
top-level `import` statements are illegal in libdefs. `Type_sig_parse` didn't handle this; it would create a sig with a global `RemoteBinding`, which hits a `failwith` in `Type_sig_pack` because it (legitimately) should never happen.

this diff makes `Type_sig_parse` ignore these invalid imports.

instead of doing this here and also in `Type_inference_js.infer_lib_file`, banning global imports should just be a parse error from the parser itself. that requires more work... we need to parse lib files differently -- which we need to do anyway because implementations should be banned. but that's left for followup work.

Reviewed By: panagosg7

Differential Revision: D26155698 (https://github.com/facebook/flow/commit/0f6d530e324ea8fcf8f48356dab0679493711c4d)

fbshipit-source-id: 0d8462ed2e5bf28c45ccd160c4790ab0a8a476e7