Refactor implicit_instantiation module in preparation for better tests
Summary:
In order to inspect and verify the results of Pierce's algorithm, I plan to add a second entry-point into the `Implicit_instantiation` module that exposes the map from tparams to their inferred types.

This diff generalizes the pinning type portion to allow callers to customize the behavior of what happens when:
1. A type can't be inferred because there aren't sufficient bounds
2. A type is inferred successfully
3. A type can't be pinned because it includes a non-Type.t
4. A type does not appear in the return type of the function

There is no change in behavior.

Reviewed By: panagosg7

Differential Revision: D26127914

fbshipit-source-id: f145327221e0699315c794a7a5c71571f9fe0ae9