[type-sig] fix exports of declared modules
Summary:
`declare module` has a (legacy?) behavior where it is implicitly a CJS module by default, unless you use `declare export ...` (except `declare export type`) to explicitly make it an ES module, or `declare module.exports: ...` to explicitly make it a CJS module.

in this mode, all values are implicitly exported. in both implicit and explicit CJS modes, types are always exported. in ES mode, both values and types require the `export` keyword.

this appears to be completely unspecified in the libdef [docs](https://flow.org/en/docs/libdefs/creation/#toc-declaring-a-module), but it's actually the original behavior before ES modules existed. most of our builtin libs are written this way.

I didn't update the docs because I'm not convinced this behavior should still exist, but the goal here is to bring type sigs up to parity with the previous behavior.

Reviewed By: panagosg7

Differential Revision: D26194233 (https://github.com/facebook/flow/commit/a3db32cc88ce2ed1922952879bedfe927cd07821)

fbshipit-source-id: 3feed07a9b0d6b61a9f1ad5734e5e63dde3ae476