functorize `merge_context_generic ` of merge_service.ml
Summary:
Before, `merge_context_generic` was used to service both the merge and
the check phase, using inputs that were components or files respectively.
Keeping track of what phase we were in was starting to become complex, since
we'd have to check `Context.phase` at various parts of the code, and use
`failwith` or "dummy" results if the result was not something we expected.

This diff fixes this by turning `merge_context_generic` into a functor `Process_unit`
that takes a `PROCESS_UNIT` module as input. The main operations that this
configuration module needs to provide is the kind of I/O of the processing and
how to process inputs given a set of dependencies (`process_with_deps`).
This last function is implemented by calling into the relevant functions of
merge_js.ml.

This refactoring removes 3 `failwith`s from merge_service.ml, and simplifies
input to "check" phase related functions to refer to a file instead of a component
of size 1.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D26206061

fbshipit-source-id: 09f105ed3c600d98650f20a29354e7f7ec6f58da