Functorize Implicit_instantiation module
Summary:
Instead of passing around functions directly and forcing every user of the Implicit_instantiation module to add code for their use case into the Implicit_instantiation module, this diff turns that module into a functor that follows the [observer pattern](https://en.wikipedia.org/wiki/Observer_pattern).

The struct passed in describes how to handle each case in Pierce's algorithm and the `run` function returns a mapping from type param names to the values those functions return.

I was able to replace both users of the module (tests and post-inference run), which got the testing-specific and post-inference-specific code out of the module.

It will be easy to make a new struct to be used during inference when we are further along in the local inference project

Reviewed By: panagosg7

Differential Revision: D26210127

fbshipit-source-id: e9b4c76484e90fce3f474ad6811e994ef993e977