don't exit when merely touching .flowconfig
Summary:
`touch .flowconfig` triggers a filesystem change, and we restart whenever we're told the flowconfig changed, without actually checking to see if the contents changed.

now we compare the hash from when we started the server to the hash after the update. we still restart if _anything_ changed, even a comment, but we have plans to work on handling changes incrementally later.

Reviewed By: nmote

Differential Revision: D26252756 (https://github.com/facebook/flow/commit/cff5f3bc637b14388ff2c89e7663eec1161825a1)

fbshipit-source-id: 48074e6e83d5d13a196e67df04345d85d3d208b4