Ban doubly-internal names
Summary:
This is going to be a natural consequence of the upcoming D26153038, which converts the name type into a variant.

I didn't think I would need to do this until I came across a code path that looks up a name in the name map, then if nothing is found, tries the internal version of that name. This means that potentially we could take an already-internal name, and make it doubly (or triply!) internal by adding another `.`. In order to support this with the new name type, we'll need a way to convert something of type `name` to an internal version of that name. Since we don't have a way to represent a doubly-internal name, I'm making the change that could potentially change behavior in this diff, so that the big one doesn't change behavior.

Hopefully this never happens, but I'd like to be more sure.

Reviewed By: panagosg7

Differential Revision: D26188208

fbshipit-source-id: 1e5d210f76dcc7235d04e2b290333c2cd28c9777