Do not transform internal names into SingletonStrT types when evaluating GetValuesT
Summary:
It's not clear that it's possible for an internal name to show up here, though if that's true, it cannot currently be proven by construction in my upcoming change D26153038.

If internal names do show up here, it's unlikely that the user intended them to be transformed into singleton types and make their way into the resulting union type.

Reviewed By: panagosg7

Differential Revision: D26255165

fbshipit-source-id: 67a4ad88dbe3d07485d7a6fab12358aa857bb7ae