interfaces exhibit proper subtyping with object literals
Summary:
Interfaces did not previously allow covariant subtyping with object literals, a difference in behavior with object types that would result in spurious errors.

This also fixes the behavior of optional properties in interfaces, which previously were *always* covariant, even when it was unsafe to be so. This uncovers a few errors.

Reviewed By: jbrown215

Differential Revision: D25984240

fbshipit-source-id: b3fb2b330c42445d47e226a3c5280ba9fbcc6d35