[fix] suggest the right kind of imports
Summary:
Fixes a bug in autoimport code actions where we'd suggest `import type` as a fix for a missing value, and vice versa.

Note: this already worked correctly for autocomplete; this does a similar thing for code actions. Autocomplete already figures out whether we're in a value or a type identifier on its own (to suggest local types or values), so we have to do a similar AST traversal on code actions.

Reviewed By: vrama628

Differential Revision: D26462493

fbshipit-source-id: c2d8ac0ac637e8c971ae7c4f594864cc1218a61b