[chore] pass kind of import declaration into import_specifier mapper
Summary:
because you can write either `import type { T }` or `import { type T }`, it's impossible to know the kind of an import specifier without knowing the kind of the import declaration itself.

the `import_specifier` visitor is mostly useless without this info; you end up having to override `import_declaration` itself which is way too coarse.

this patch passes the declaration's kind down to the specifiers in `Flow_polymorphic_ast_mapper`. we could do the same for `Flow_ast_mapper` but I don't need it right now and it has a bunch of callers.

Reviewed By: panagosg7

Differential Revision: D26424468 (https://github.com/facebook/flow/commit/f153260dfe2e0ea4719bb90bc4e515187c6b56ef)

fbshipit-source-id: a82d4bfdf7b0090f63a248f5f3c3b177469e1d42