fix get-def jumping through annotations instead of to them
Summary:
if Go to Definition on an identifier leads to an import statement, we jump through the import. however, it currently jumps all the way to the type's `def_loc`, which is not the value being imported, it's the definition of that value's type (aka Go to Type Definition!).

if the imported value has an annotation in the exported file, we should jump there so you can see the value. get-def on that annotation will take you to the type definition.

Changelog:
- Fixed a bug in Go to Definition on imported values that skipped all the way to the value's type definition instead of the value's definition.

Reviewed By: vrama628

Differential Revision: D26403592 (https://github.com/facebook/flow/commit/272e5bba5c8ffd0395f47b4e40bd16b1ac1561d3)

fbshipit-source-id: 18e1acdb0936000c4b442f7fcc631a2a10b7dc9f