Inject behavior for visitors
Summary:
This diff continues the change from the previous diff in this stack. In addition
to the dependencies field, now all of the visitors are functions instead of the
concrete "node" data type.

In types-first 2.0 mode, instead of eagerly visiting all reachable nodes during
merge, as we currently do in new signatures mode, we want to be lazy. By
parameterizing the Type_sig_merge module in this way, it is possible to re-use
this code in both eager and lazy contexts.

Reviewed By: panagosg7

Differential Revision: D26479747

fbshipit-source-id: f5a50cb2975b45dd3c6892dc181430b31e6ab93f