Inject behavior for merging dependencies
Summary:
This changes the "dependencies" fields of file record from a concrete data type
to a function. The function is provided by the callers, currently merge_service
under new_signatures mode.

In new_signatures mode, we call into this function during merge, where we have
the entire component loaded. By injecting the behavior instead, this change
enables re-using this module in full types-first 2.0, where resolving a
dependency looks quite different.

As a bonus, since the passed-in function hides the recusive nature of the
component, we don't need to thread the component array through every function
call in Type_sig_merge.

Reviewed By: panagosg7

Differential Revision: D26479746

fbshipit-source-id: 907d8943eef4208d82b97d71a072e9f910cf83b3