Pass aloc tables to Context.make
Summary:
Before this change, callers needed to create a filename map of (lazy) aloc
tables before calling make_ccx, then pass in a rev_table to make.

Instead, let's pass the file's aloc_table to make and hide the logic for
creating a reverse table.

I find this cleaner as-is, but it's also helpful for types-first 2.0 where we
create contexts as-needed for dependencies, but might associate them with an
existing component context in the case of cycles.

That is, I don't know the full set of files in a given component ahead of time,
so I can't create the aloc table map ahead of time. I originally had an
`add_aloc_table` API for this purpose, but I find this much cleaner.

Reviewed By: nmote

Differential Revision: D26470258

fbshipit-source-id: 70fb8eed0daf4b3dbb35bc447ac51bebd1d7baa0