Convert mutable Context.sig_t into immutable Type.TypeContext.t
Summary:
The Context.sig_t record contains information that is often necessary to
"interpret" the meaning of some type. For example, if the type is an OpenT, we
must interrogate the graph to see whether it is resolved, and if so, to what.
Alternatively, if the type is an ObjT, we might want to see what its properties
are, which are stored on the side in Context.sig_t.

This diff moves the definition of this record from the Context module into the
Type module. The actual records are still stored in the context, but updated a
bit differently.

The following diff, which adds a SigT type, will make use of this relocation to
provide a new kind of type which includes a `TypeContext.t -> TypeContext.t`
function.

Reviewed By: jbrown215

Differential Revision: D26480155

fbshipit-source-id: 5a6c438da9107b1e419e490b5b803bd0530e17c2