more precise refinement invalidation
Summary:
It is common for a variable to pass from an "uninitialized" state to an "initialized" state, and be written by closures in such a way that it remains in that "initialized" state. Previously, in such cases we would still fall back to the most general type for such a variable on function calls, which would include the "uninitialized" state.

This diff does a bit more work to collect just the writes that are done by closures, and on calls, widen the environment with these writes. In particular, this fixes the precision issue above and removes a significant class of false positives.

We have to be a bit careful about performance here, since the collection and widening steps both introduce tvars and it's easy to imagine cases that can cause too many useless flows through this indirection. The usual check `specific == general` applies, but we also prevent the common case where a widened variable can loop back and cause further widening. Specifically, if `t` is the tvar collecting writes of a variable `x` by closures and `X` is the state of `x` before a call, we might create a tvar `w` to represent `t | X`, but if `w` comes back as `X` then we don't create a new `w'` to represent `w | X`, since we can just shortcut that back to `w`. Both these optimizations are deemed critical in that they fire lots of times; other optimizations are possible but don't fire as much, and instead cause the overall operation to become too heavyweight.

Reviewed By: panagosg7

Differential Revision: D14018024

fbshipit-source-id: b4bde65a1b705e29957ebccebdf81216f608fe57