fix buggy constraint-based refinement invalidation with lexical analysis
Summary:
Previously, a function `f` that wrote to a captured variable `x` would include that variable in the changeset of its function type; moreover, a call would include the calling environment in its call type. Then, if `f` flowed to the call and the calling environment included `x`, then `x` would be "havoc"-d, meaning that we would forget its flow-sensitive state and fall back to its most general type, reflecting the possibility that the state after the call might not be the same before the call.

While this analysis aimed to be precise, it was also buggy in a number of ways:
1. If `f` instead called `g` and it was `g` that wrote to `x`, then we would sometimes not include `x` in the changeset of `f`. Doing this correctly would need a major rewrite of changeset propagation involving "effects," of the kind described in our OOPSLA 2017 paper, and it's not clear that it would be performant.
2. If `f` went through an annotation before it was called, then we would also not include `x` in the changeset, because annotations have empty changesets; again, fixing this would need a major rewrite, with function type annotations accumulating the changesets of function types that flowed to them.

This diff cuts the dependency of refinement invalidation on this kind of constraint-based analysis. Instead, it considers a variable `x` that is ever written by *any* closure that captures it as a "havoc"-able variable, and invalidates the state of all havoc-able variables in the environment at every call. This is in essence a coarse approximation of the constraint-based analysis, but in practice works well enough because there are not too many such variables. On the other hand, it is immune to the kind of bugs described above, and longer term, it unblocks the reuse of `Ssa_builder` in modeling all of control-flow analysis (including the treatment of conditionals, loops, and abnormal control flows), so that all environment state management can be moved out of `statement_js.ml`.

In a later diff, we make a further improvement to this analysis that gets rid of a common class of false positives.

Fixes #7541

Reviewed By: panagosg7

Differential Revision: D13962715

fbshipit-source-id: b10a5a590da029e9ac255d45c4af788d7c3879a8