Strict ES6 import/export lint error handles default re-exports
Summary:
The strict ES6 import/export lint errors include an error for when an export is explicitly named "default" instead of using a default export declaration. This also currently applies to re-exports renamed "default" (`export {remote as default} from 'other'`) but due to the current implementation does not handle `export {default} from 'other'`.

Let's make sure to error on re-exports of a default export, suggesting that they be rewritten as an `import` followed by a `default export` instead. The error message should also be tweaked in the case of a re-export renamed to "default", as the suggestion for how to fix is different than for a non-re-export. This required adding a new `is_reexport` property to the `EExportRenamedDefault` error so we can keep track of both the original name (if one exists) and if it is for a re-export so that the error message can be constructed appropriately.

Reviewed By: pieterv

Differential Revision: D26556326

fbshipit-source-id: 1a9125a46ac96f2d77a0d370f8b28435a2a2228c