Drop unhelpful MakeSendErrors profiling
Summary:
What this profiling is measuring is just the creation of the intermediate result
callback, but all the computation happens inside of the callback. What we would
prefer is to add the profiling inside of the callback itself, but this is
unfortunately not possible. The callback is called from inside the merge
callback passed to MultiWorkerLwt.call, which is not a Lwt context.

This diff just removes the profiling entirely. It might be nice for future work
to enable calling into Lwt from merge. However, we already have a good sense of
the performance implications of MultiWorker jobs through other profiling.

Reviewed By: nmote

Differential Revision: D26648373

fbshipit-source-id: 12496475a1a6a2d18f042bb934f752a58ddf0d09