Include sighashes in saved state
Summary:
This diff includes the sighash for each file in the saved state, loads the sighash from saved state if the flag is turned on, and makes the changes needed to take advantage of those loaded sighashes for recheck optimizations.

## Implementation

This change is made up of two main components:
* First, there is the machinery to save the sig hashes to saved state and then load them again (on the condition that the appropriate flowconfig flag is set). This component makes up the bulk of the changes here, but is very simple.
* Second, there are the changes to `Merge_stream` and `Context_heaps` that change the merging logic so that we can actually take advantage of these loaded sighashes. These changes are more interesting, and deserve special attention.

I'll focus on this second component here.

At first blush, it may seem like it is be possible to simply load in the sighashes from saved state, and allow the existing machinery to skip merging and checking whenever the newly-computed sighashes match the old, loaded ones. Unfortunately, this is a fantasy.

The issue is that we may skip merging a component because none of its dependencies changed, which would mean that we still don't have a sig context for it. Then, we may later need to merge or check a downstream component (for example, if another of its dependencies has changed), and to do that we'll need the missing sig context.

To address this issue, it would be possible to carefully track every component that has a missing sig context, and backtrack if it is needed for a downstream file. I may still implement such logic in the future. However, for now I think it makes sense to simply merge every file with a missing sig context, but use the newly-available sighashes to drive skipping in the check phase. This is the approach that I have taken in this diff.

Previously, we didn't distinguish between newly-added sig contexts and those that changed. Now, however, we have the ability to know that a sig context has not changed (based on the sighashes loaded from saved state) even when it is new. So, we need to add a case for that -- if the sighashes are the same, but there wasn't an old sig context, we need to write it, but not add it to the `sig_new_or_changed` set. This is what the changes in `Merge_stream` and `Context_heaps` do.

So, on the initial recheck after a lazy init, we will merge all of the files that we consider for merge, just like we do today. However, we will use the sighashes to gather a set of files that have actually changed, and use that set to determine which files we can skip in the check phase. This should, in theory, reduce the number of files we check in the check phase of an initial recheck after a lazy init, thereby reducing the time spent in these early rechecks.

## Caveats

* There are some performance ramifications (see below).
* There are some serious hash stability issues that hamper the effectiveness of this optimization:
  * When testing on real-world inputs, I found that ~40% of files had sighashes that were unstable from run-to-run. They appear to be stable for the life of a given Flow server. I have not investigated this further. I believe that the hashes computed by types-first 2.0 will be more stable, so I am going to put off a detailed investigation in the hopes that this resolves itself.
  * The hashes include the absolute file paths of various source files, so they will not be stable from machine-to-machine or repo-to-repo unless the absolute path to the repo is identical across invocations. This assumption is unlikely to hold in practice. I plan to address this issue next.
* This may cause us to miss some already-committed errors, similar to how we already can miss already-committed errors in lazy mode. However, the extent to which this is true would increase: today, if you modify a file, Flow will display errors for all of its dependents after the initial lazy recheck. With this change, Flow would only display errors for dependents if your changes might have actually affected those dependents. This could be a plus or a minus depending on your point of view.
* As part of the `flow status` output, when Flow is running in lazy mode, we report the number of files that are currently being checked. Without additional changes, we will now report the number of files that we have at one point considered for checking, even if we ended up skipping them. This might merit some additional thought, but it's unlikely to be a noticeable difference for users in practice.

## Performance

This slows down startup time when turned on (but only barely). It does not appear to affect startup time when turned off. It also slightly increases the size of the saved state blob.

When testing on a benchmark recheck, it was only able to skip checking for less than 1% of the files on an initial recheck, due to the hash stability issues mentioned above. However, tests pass, and this has no measurable performance regression when it's turned off, so I would like to land it and then work on improving hash stability later.

Once the hash stability issues are worked out, we should be able to skip as many files in the check phase of an initial lazy recheck as we would for any other recheck. I believe this will be a significant performance improvement, though we will still have to pay the cost of parsing and merging all of the files. Fortunately, the cost of merging will go down dramatically with types-first 2.0, and the check phase is the most costly even today.

Reviewed By: samwgoldman

Differential Revision: D25137486

fbshipit-source-id: 3e9e5846baec287c880211b13171629668457dde