Tighten up the representation of merge and check results
Summary:
Now that classic is gone, this change is possible. In types-first, merge never
returns errors, warnings, or coverage. This diff changes the merge job to return
only suppressions and the duration.

Note that we need suppressions from merge in order to stream results during
check. Otherwise we don't need suppressions from merge.

Currently it is possible to pass unchecked files to merge and check. Before this
change, we would return empty data. This diff changes this to return None
instead of Some data. I simply prefer this, but hopefully in the future we can
ensure that we don't pass unchecked files to merge/check in the first place, in
which case this will be unnecessary.

The impetus for this change is that I want to return an additional bit of
information from merge, specifically whether the merged file is changed (context
opts). This information doesn't apply to check results, but merge results and
check results were conflated (necessarily for classic, but no longer required).
In the process of separating the types for merge/check results, I realized that
I could make the changes here.

I will add the additional information mentioned above in follow-up work.

Reviewed By: panagosg7

Differential Revision: D26697824

fbshipit-source-id: 7aae638f47a73f95cd9d6a7e94fe25ad6293871c