Don't accumulate suppressions while streaming errors
Summary:
When classic and types-first coexisted, the streaming error callback needed to
account for two cases:

1. Errors found during merge in classic
2. Errors found during check in types-first

In the former case, we accumulated suppressions, relying on an invariant that
any errors we found, if suppressed, were suppressed by a comment in that file or
a dependency of that file.

In the latter case, since check does not run in dependency order, we need to
have all the suppressions already. For this case, we rely on the suppressions
passed to mk_intermeidate_result_callback, which is accumulated during merge.

This diff deletes the code added to handle case (1) which is no longer relevant.

Reviewed By: panagosg7

Differential Revision: D26697822

fbshipit-source-id: 9bfc1f22a380aa7f9f7ae734cd49b3d3b3646b93