Have merge jobs return diff status instead of querying in merge stream
Summary:
As we merge components, we want to know whether the just merged component has
changed -- that is, whether it's exports are observably different from the last
time we merged. We use this information to skip files whose dependencies have
not changed. This is called "recheck optimizations" in Flow's architecture.

Before this diff, the merge stream module, which schedules components for merge
in dependency order and deals with skipping, would query for the changed status
directly from the heap.

This diff moves the responsibility into the merge job, which will determine the
status in the worker and pass that information back to the merge stream in the
form of a boolean.

This change should be behavior preserving, although it does open up some
interesting questions about the existing behavior. For example, is it actually
correct to return "false" from the case where checked = false? What if the files
in that component were previously checked?

This change is a step toward making the merge stream module more general, and
suitable for use in types-first 2.0.

Reviewed By: panagosg7

Differential Revision: D26702268

fbshipit-source-id: 6b0ff92266f8ed4259c94d18967335344975b7d6