fix race in dir_coverage test
Summary:
dir_coverage has been failing fairly repeatably on mac. my theory is that there's a race between the `rm` and the file watcher noticing it before the `batch-coverage` call, so `--wait-for-recheck` doesn't help because no recheck has been started yet.

`force-recheck` is made for this kind of problem and should work on deleted files

Reviewed By: jbrown215

Differential Revision: D26778956 (https://github.com/facebook/flow/commit/10e9ea15b0631f5e8d8a85f6a2ef2627da13a7a9)

fbshipit-source-id: e3322ef0a772835eb3dc834b9df4b36caba5a8fd