Fix types-first signature builder handling of dynamic import
Summary:
Before this fix, types-first would convert a dynamic import a non-equivalent
signature AST. A dynamic import returns a promise which resolves to the
namespace object of the referenced module, but types-first would return the
namespace object directly.

For example, the module:

```
module.exports = import("foo");
```

used to be (incorrectly) converted into:

```
import * as $1 from "foo";
declare module.exports: typeof $1;
```

and is now (correctly) converted into:

```
import * as $1 from "foo";
declare module.exports: Promise<typeof $1>;
```

The code changes for this are gnarly. It might be better to factor out some of
the logic to create Promise typeapps, e.g., but this code is going to be deleted
very soon.

Fixing this for types-first so I can add a test and then make that test pass for
types-first 2.0, which will have a much nicer implementation.

Reviewed By: panagosg7

Differential Revision: D26756244 (https://github.com/facebook/flow/commit/b3177869bce59d68baa42b405b529c090646ed8d)

fbshipit-source-id: 6848693c30ab950be1537dd905f94281568d7f7e