Incremental mark and sweep
Summary:
This diff implements an incremental mark and sweep pass for shared memory, which
is used to find free space and trigger compaction.

Before this diff, we triggered compaction by counting the "wasted space" when a
heap entry was deleted. This method does not work when heap entries have
references between them, however.

For example, consider a heap with objects A and B, where A points to B, and only
A is references from the hash table. If we remove A from the hash table, we've
also made B unreachable. Without this change, we would only account for the free
space in A.

In this diff, we do a full heap scan, marking live objects reachable from the
root set, to find garbage. This is more expensive than counting free space on
delete, but enables us to write heap objects that keep other heap objects alive.

To deal with the expense of a heap scan, we only do this work during "slack
time" between rechecks. We furthermore incrementalize the work, so that a
request can suspend the scan, and let the server resume after sending a
response.

The finer points of the GC are documented inline.

Reviewed By: nmote

Differential Revision: D25447328

fbshipit-source-id: 8ffd59bd06b10e9618d8693b6e5d8f2de9bc8df1