[indexed-access] Type normalizer support for indexed access
Summary:
Add custom behaviour rather than just piggy-backing off of `$ElementType<...>`. This will be required anyway to support optional indexed access, which does not have a legacy analog.

In the future, we will eliminate `$ElementType<...>` and rename/replace `ElementType` in the code with `IndexedAccessType`, and remove the `is_indexed_access`.

Reviewed By: panagosg7

Differential Revision: D26891740

fbshipit-source-id: 633824572ec619b4f450c07960164a39ae9d108a