Add Builtins module
Summary:
This module will be used to represent builtins instead of using an unsealed object.

It allows deferred writes via the `on_missing` callback to `get_builtin`.
It also provides a way to validate that every read has a corresponding write. In D26488590 we use this function to emit errors on missing builtins.

Reviewed By: panagosg7

Differential Revision: D26488592

fbshipit-source-id: c22d500c5f419e48c1771078e1e4ea61cda0748b