Enable new signatures by default
Summary:
This diff switches the default for the flowconfig, so that new signatures
mode is enabled by default. Users can opt out of this behavior by explicitly
setting `experimental.new_signatures=false` in the flowconfig.

My hope is to swap the default in the Flow v0.147, then remove the opt-out in
v0.148, so that everyone is on the new signatures infra and we can delete the
old signature verifier/generator code.

Reviewed By: panagosg7

Differential Revision: D26923323

fbshipit-source-id: 0d3518787d27faabfef7368d1e2e82409a1fecdb