Fix unification by rank
Summary:
When we call `goto`, the order of root1 and root2 is significant. We've already
determined that we want to make id1 a Goto node pointing to root2 by comparing
the ranks of root1 and root2.

For the case where root1 is resolved and root2 is unresolved, we also want to
preserve the resolved-ness. We do this by replacing id2 with a resolved root,
but still making id1 point to id2.

I broke this in D5540929 (https://github.com/facebook/flow/commit/317faa89610c3d8b6416780c2dc7a101a029597b). You can see in the before part of those changes that
we did the correct thing before (change root2, make id1 point to root2) and the
wrong thing after (make id2 point to root1).

In practice, I don't know that we'll see any perf win from this, but it's worth
at least being correct with respect to the underlying algorithm!

Reviewed By: panagosg7

Differential Revision: D26957428

fbshipit-source-id: 2fde7316ac598d35424edc97c7be49391e22b644