[lsp] more specific exception when editing unopened file
Summary:
the LSP requires that a client send a `didOpen` before `didChange`. we rely on this invariant to assume that the changed file is already in `open_files`.

when that's not the case, we get a generic `Not_found` exception. this new exception tells us what is not found.

in either case, we weren't/aren't catching the exception. it bubbles up to the LSP handler and gets caught and logged. since it's a notification, no response to the client is necessary (or allowed).

Reviewed By: nmote

Differential Revision: D26983795

fbshipit-source-id: dee1dd51c79f56be9039fc9f97a07378d722095e