remove all=true from newtests
Summary:
this is a non-default setting, so it should not default to true for all tests. tests that want to test that setting should opt into it.

in particular, this is masking a bad test in `lsp/queries` that expects coverage to not return results for non-@flow files. will fix that next in the stack.

Reviewed By: samwgoldman

Differential Revision: D27012503

fbshipit-source-id: f2bf8e7fd4eeb647751655e6ee0eaf6cd07bd74a