[lsp] more specific exception when editing unopened file
Summary:
the LSP requires that a client send a `didOpen` before `didChange`. we rely on this invariant to assume that the changed file is already in `open_files`.

when that's not the case, we get a generic `Not_found` exception. this new exception tells us what is not found.

in either case, we weren't/aren't catching the exception. it bubbles up to the LSP handler and gets caught and logged. since it's a notification, no response to the client is necessary (or allowed).

Reviewed By: nmote

Differential Revision: D26983795 (https://github.com/facebook/flow/commit/e23278bc17e6a0b5a2c52719d24b6bc5bb716931)

fbshipit-source-id: 13e2a8aba376852b1b093243afd5003803a480b4