[refactor] move abstract_locations default out of flowConfig
Summary:
when we store defaults in `flowConfig.ml`, we can't tell whether the user explicitly set the value to the default, or it was unset. this makes it hard to reformat the flowconfig.

instead of storing `bool`, we should store `bool option`, and move the default to `commandUtils.ml` where we construct `Options` by combining the CLI flags, flowconfig and defaults.

Reviewed By: gkz

Differential Revision: D27047153

fbshipit-source-id: 457d14a49a0ebf7c128a0d479d0ccc502724d3cd